Rename and re-arrange form component.
